requirements.txt
Pillow==8.3.2
markovify